Rename package to match directory, and run make goporto
