+++ pkg/stanza/internal/fileconsumer/finder_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package file
+package fileconsumer
 
 import (
 	"io/ioutil"
